Sort with KDF.
There are issues with bucketing signed integers, because
for smaller values, the top bits are now either
1000... or 0111... meaning we'll fall into mostly two buckets.

The solution to this seems to be a more complicated extra-shift
scheme that can account for negative numbers, but it may not
be worth it.

No doubt floats will have the same issue.