Relbits is an experiment in key compaction.
Basically we can scan the input data and use PEXT to compress
the keys such that irrelevant bits -- those that does not influence
the relative order of all the values if they are removed -- can
be ignored.

This has applications in radix sorting and LUTed binary search
which also relies on bucketing.